requirements/requirements.txt
gurux-dlms
beautifulsoup4
paho-mqtt
lxml
pyserial
cryptography==37.0.2
datetime
setuptools
python-dotenv
logging
traceback
datetime
pathlib
os